+++ Common/src/main/java/net/merchantpug/apugli/action/factory/entity/AddVelocityAction.java
@@ -1,7 +1,6 @@
 package net.merchantpug.apugli.action.factory.entity;
 
 import com.mojang.math.Vector3f;
-import io.github.apace100.apoli.Apoli;
 import io.github.apace100.apoli.data.ApoliDataTypes;
 import io.github.apace100.apoli.util.Space;
 import io.github.apace100.calio.data.SerializableData;
@@ -11,12 +10,11 @@ import net.merchantpug.apugli.platform.Services;
 import net.minecraft.util.Mth;
 import net.minecraft.world.entity.Entity;
 import net.minecraft.world.entity.player.Player;
-import org.apache.logging.log4j.util.TriConsumer;
 
 import java.util.ArrayList;
 import java.util.List;
 
-public class ClampedAddVelocityAction implements IActionFactory<Entity> {
+public class AddVelocityAction implements IActionFactory<Entity> {
 
     @Override
     public SerializableData getSerializableData() {
@@ -28,10 +26,10 @@ public class ClampedAddVelocityAction implements IActionFactory<Entity> {
                 .add("horizontal_modifiers", Services.PLATFORM.getModifiersDataType(), null)
                 .add("vertical_modifier", Services.PLATFORM.getModifierDataType(), null)
                 .add("vertical_modifiers", Services.PLATFORM.getModifiersDataType(), null)
-                .add("horizontal_clamp_modifier", Services.PLATFORM.getModifierDataType(), null)
-                .add("horizontal_clamp_modifiers", Services.PLATFORM.getModifiersDataType(), null)
-                .add("vertical_clamp_modifier", Services.PLATFORM.getModifierDataType(), null)
-                .add("vertical_clamp_modifiers", Services.PLATFORM.getModifiersDataType(), null)
+                .add("horizontal_post_modifier", Services.PLATFORM.getModifierDataType(), null)
+                .add("horizontal_post_modifiers", Services.PLATFORM.getModifiersDataType(), null)
+                .add("vertical_post_modifier", Services.PLATFORM.getModifierDataType(), null)
+                .add("vertical_post_modifiers", Services.PLATFORM.getModifiersDataType(), null)
                 .add("space", ApoliDataTypes.SPACE, Space.WORLD)
                 .add("client", SerializableDataTypes.BOOLEAN, true)
                 .add("server", SerializableDataTypes.BOOLEAN, true)
@@ -50,22 +48,24 @@ public class ClampedAddVelocityAction implements IActionFactory<Entity> {
                 data.getFloat("y"),
                 data.get("z"));
         space.toGlobal(vec, entity);
-        vec.set((float) Services.PLATFORM.applyModifiers(entity, getModifiers(data, "horizontal_modifier"), vec.x()), (float) Services.PLATFORM.applyModifiers(entity, getModifiers(data, "vertical_modifier"), vec.y()), (float) Services.PLATFORM.applyModifiers(entity, getModifiers(data, "horizontal_modifier"), vec.z()));
+        vec.set(applyModifiers(data, entity, vec.x(), "horizontal_modifier"),
+                applyModifiers(data, entity, vec.y(), "vertical_modifier"),
+                applyModifiers(data, entity, vec.z(), "horizontal_modifier"));
         if (!data.getBoolean("set")) {
             vec.add((float) entity.getDeltaMovement().x(), (float) entity.getDeltaMovement().y(), (float) entity.getDeltaMovement().z());
         }
         entity.setDeltaMovement(
-                clampModifiers(data, entity, vec.x(), "horizontal_clamp_modifier"),
-                clampModifiers(data, entity, vec.y(), "vertical_clamp_modifier"),
-                clampModifiers(data, entity, vec.z(), "horizontal_clamp_modifier"));
+                applyModifiers(data, entity, vec.x(), "horizontal_post_modifier"),
+                applyModifiers(data, entity, vec.y(), "vertical_post_modifier"),
+                applyModifiers(data, entity, vec.z(), "horizontal_post_modifier"));
         entity.hurtMarked = true;
     }
 
-    private float clampModifiers(SerializableData.Instance data, Entity entity, float original, String modifierKey) {
+    private float applyModifiers(SerializableData.Instance data, Entity entity, float original, String modifierKey) {
         if (!data.isPresent(modifierKey) && !data.isPresent(modifierKey + "s")) {
             return original;
         }
-        return Mth.clamp(original, -Mth.abs((float) Services.PLATFORM.applyModifiers(entity, getModifiers(data, modifierKey), original)), Mth.abs((float) Services.PLATFORM.applyModifiers(entity, getModifiers(data, modifierKey), original)));
+        return ((float) Services.PLATFORM.applyModifiers(entity, getModifiers(data, modifierKey), original));
     }
 
     private <M> List<M> getModifiers(SerializableData.Instance data, String modifierKey) {