+++ Common/src/main/java/net/merchantpug/apugli/registry/action/ApugliEntityActions.java
@@ -1,6 +1,7 @@
 package net.merchantpug.apugli.registry.action;
 
 import net.merchantpug.apugli.action.factory.IActionFactory;
+import net.merchantpug.apugli.action.factory.entity.meta.PacketAction;
 import net.merchantpug.apugli.platform.Services;
 import net.merchantpug.apugli.action.factory.entity.*;
 import net.minecraft.world.entity.Entity;
@@ -8,12 +9,13 @@ import net.minecraft.world.entity.Entity;
 public class ApugliEntityActions {
     
     public static void registerAll() {
-        register("clamped_add_velocity", new ClampedAddVelocityAction());
+        register("add_velocity", new AddVelocityAction());
         register("explode", new ExplodeAction());
         register("fire_projectile", new FireProjectileAction());
         register("raycast", new RaycastAction());
         register("resource_transfer", new ResourceTransferAction());
-        register("rocket_jump_raycast", new RocketJumpRaycastAction());
+        register("explosion_raycast", new ExplosionRaycastAction());
+        register("packet", new PacketAction());
         register("set_no_gravity", new SetNoGravityAction());
         register("spawn_item", new SpawnItemAction());
         register("spawn_particles", new SpawnParticlesAction());