+++ Common/src/main/java/net/merchantpug/apugli/util/Side.java
@@ -1,0 +1,5 @@
+package net.merchantpug.apugli.util;
+
+public enum Side {
+        CLIENT, SERVER
+}